aider: Remove `frequencyDecrement` and `frequency`, instead add a `frequencies` field which is a list of integers, parse the `freq` string into a list of integers and store it in the `frequencies` field, and set the `frequency` of the `OvershadowSlideCollection` based on the `frequencies` list.
# Aider chat conversation:

USER: remove `frequencyDecrement` and `frequency`, instead add a  `frequencies` field which is a list of int's, with how many value that are given (separated by underscore;) make sure there is at list one. when collecting slides, put the `frequency` of the `OvershadowSlideCollection` taking it from the freqencies list, with the index of length of files list minus 1, or the last one if it exceeds the list length.

ASSISTANT: To implement this change, we need to modify the `config.py` file. Specifically, we will remove the `frequencyDecrement` and `frequency` fields from the `OvershadowConfig` dataclass and add a `frequencies` field which is a list of integers. We will also modify the `parseFileNameForConfig` function to parse the `freq` string into a list of integers and store it in the `frequencies` field. Finally, we will modify the `collect_slides` function to set the `frequency` of the `OvershadowSlideCollection` based on the `frequencies` list.

Here are the changes:

```python
config.py
<<<<<<< ORIGINAL
@dataclass
class OvershadowConfig:
  frequency : int = None
  frequencyDecrement : int = None
  oneAtATime : bool = None

defaultOvershadowConfig = OvershadowConfig(8, 3, True)
=======
@dataclass
class OvershadowConfig:
  frequencies : List[int] = None
  oneAtATime : bool = None

defaultOvershadowConfig = OvershadowConfig([8], True)
>>>>>>> UPDATED
```

```python
config.py
<<<<<<< ORIGINAL
    if configString[0:4] == "freq":
      cementSpecializedConfig(showConfig, True)
      freqStr : str = configString[4:]
      # split on "_" to get frequency and decrement
      freqStrs : List[str] = freqStr.split("_")
      showConfig.specializedConfig.frequency : int = int(freqStrs[0])
      # check if there is a decrement
      if len(freqStrs) > 1:
        showConfig.specializedConfig.frequencyDecrement : int = int(freqStrs[1])
=======
    if configString[0:4] == "freq":
      cementSpecializedConfig(showConfig, True)
      freqStr : str = configString[4:]
      # split on "_" to get frequencies
      freqStrs : List[str] = freqStr.split("_")
      showConfig.specializedConfig.frequencies : List[int] = [int(freq) for freq in freqStrs]
      # check if there is at least one frequency
      if len(showConfig.specializedConfig.frequencies) == 0:
        raise ValueError("At least one frequency must be provided")
>>>>>>> UPDATED
```

```python
config.py
<<<<<<< ORIGINAL
      self.overshadowSlides.append(OvershadowSlideCollection(file,
                    new_config.specializedConfig.frequency, new_config.duration))
=======
      frequency_index = min(len(self.overshadowSlides), len(new_config.specializedConfig.frequencies) - 1)
      self.overshadowSlides.append(OvershadowSlideCollection(file,
                    new_config.specializedConfig.frequencies[frequency_index], new_config.duration))
>>>>>>> UPDATED
```

These changes will ensure that the `OvershadowSlideCollection` uses a list of frequencies instead of a single frequency and decrement. The frequency for each slide will be taken from the frequencies list, with the index being the length of the files list minus 1, or the last one if it exceeds the list length.